ACS 2.75
Problem with the email sending is somewhat resolved: It works fine on
tablet, and would probably work fine on anyone's phone. The reason the
emulator was encountering errors was because there was no email sending
app installed on it... In any case, the code has been surrounded with a
try-catch block.